requirements.txt
absl-py==0.6.1
astor==0.7.1
atomicwrites==1.2.1
attrs==18.2.0
bleach==1.5.0
cycler==0.10.0
Cython>=0.29.1
gast==0.2.0
graphviz==0.8
grpcio==1.17.0
h5py==2.8.0
html5lib==0.9999999
kiwisolver==1.0.1
lxml==4.2.5
Markdown==3.0.1
matplotlib==3.0.2
more-itertools==4.3.0
mpmath==1.0.0
numpy==1.15.4
omrdatasettools==0.13
opencv-python==3.4.4.19
pandas==0.23.4
Pillow==5.3.0
pluggy==0.8.0
protobuf==3.6.1
py==1.7.0
pydot==1.4.0
pyparsing==2.3.0
pytest==4.0.1
python-dateutil==2.7.5
pytz==2018.7
requests==2.21.0
scikit-learn==0.20.1
scipy==1.1.0
six==1.11.0
sympy==1.3
termcolor==1.1.0
tqdm==4.28.1
tensorflow==1.13
Werkzeug==0.14.1
